requirements.txt
langchain 
streamlit
faiss-cpu
langchain_google_genai
python-dotenv